+++ nostr/nostr-client/src/main/java/org/tbk/nostr/client/NostrClientService.java
@@ -5,19 +5,25 @@ import lombok.Value;
 import org.tbk.nostr.base.SubscriptionId;
 import org.tbk.nostr.proto.Event;
 import org.tbk.nostr.proto.ReqRequest;
+import org.tbk.nostr.proto.Response;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.time.Duration;
 
 public interface NostrClientService {
+
     default Flux<Event> subscribe(ReqRequest req) {
         return this.subscribe(req, SubscribeOptions.defaultOptions());
     }
 
     Flux<Event> subscribe(ReqRequest req, SubscribeOptions options);
 
-    Flux<Event> attachTo(SubscriptionId id);
+    Flux<Response> attach();
+
+    Flux<Response> attachTo(SubscriptionId id);
+
+    Flux<Event> attachToEvents(SubscriptionId id);
 
     Mono<Void> close(SubscriptionId id);