+++ changelog.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Breaking
-- SimpleNostrTemplate: change signature of `sendPlain` method
+- NostrTemplate: change signature of method `sendPlain`
+- NostrClientService: change return value of method `attachTo`
 
 ## [0.0.2] - 2024-11-01